// src/app/model/telecel-subscriber-info.ts
export interface TelecelSubscriberInfo {
        id: number;
        subscriberNumber: string;
        name: string;
        dob: Date;
        idType: string;
        idNumber: string;
        gps: string;
        nextOfKin: string;
        agentName: string;
        agentNumber: string;
        simRegistrationDate: Date;
        telecelCashTransactions: string;

}